refactor: extract ui card componenent
